Implementors of SchemaRule moved to correct package
