+++ community/consistency-check/src/main/java/org/neo4j/consistency/checking/SchemaRuleContent.java
@@ -19,7 +19,7 @@
  */
 package org.neo4j.consistency.checking;
 
-import org.neo4j.kernel.impl.store.UniquePropertyConstraintRule;
+import org.neo4j.kernel.impl.store.record.UniquePropertyConstraintRule;
 import org.neo4j.kernel.impl.store.record.IndexRule;
 import org.neo4j.kernel.impl.store.record.SchemaRule;
